backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.100
redis==5.0.4
apscheduler==3.10.4
line-bot-sdk==3.11.0
qrcode[pil]==7.4.2
openpyxl==3.1.2
pandas==2.2.2
pytz==2024.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
anyio==4.3.0
greenlet==3.0.3
pillow==10.3.0
